% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{cea}
\alias{cea}
\alias{get_cekm_cea}
\title{Cumulative effects assessments}
\usage{
cea(
  drivers,
  vc,
  sensitivity,
  exportAs = "SpatRaster",
  align = "error",
  template = NULL,
  cores = NULL,
  filename = NULL,
  engine = "matrix"
)

get_cekm_cea(dat, vc)
}
\arguments{
\item{drivers}{distribution and intensity of environmental drivers as stars object}

\item{vc}{distribution of valued components as stars object}

\item{sensitivity}{matrix of environmental drivers and valued component, with same name as those used in \code{drivers} and \code{vc}}

\item{exportAs}{string, "SpatRaster".}

\item{align}{alignment policy, one of "error", "reproject", "template".}

\item{template}{optional SpatRaster to align to when \code{align = "template"}.}

\item{cores}{number of threads for terra::app.}

\item{filename}{optional path to write the output raster stack (useful for large jobs).}

\item{engine}{calculation engine, one of "matrix" (fast, in-memory) or "terra" (streaming/chunked).}

\item{dat}{TODO}
}
\description{
Assessment of cumulative effects using the Halpern et al. 2008 method.
}
\section{Functions}{
\itemize{
\item \code{get_cekm_cea()}: get effects per km2

}}
\examples{
drv_paths <- system.file(
  "extdata/rasters",
  c("pressure_shipping.tif", "pressure_climate.tif"),
  package = "rcea"
)
vc_paths <- system.file(
  "extdata/rasters",
  c("vc_cod.tif", "vc_salmon.tif"),
  package = "rcea"
)
drivers <- terra::rast(drv_paths)
names(drivers) <- c("shipping", "climate")
vc <- terra::rast(vc_paths)
names(vc) <- c("cod", "salmon")
sens <- matrix(
  c(
    0.8, 0.5,
    0.2, 0.7
  ),
  nrow = 2,
  dimnames = list(c("cod", "salmon"), c("shipping", "climate"))
)
ce <- cea(drivers, vc, sens, exportAs = "SpatRaster")
ce

}
