% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumul.R
\name{cumul}
\alias{cumul}
\title{Cumulative footprint}
\usage{
cumul(dat)
}
\arguments{
\item{dat}{list of stars or terra objects to measure cumulative footprints}
}
\description{
Function to evaluate cumulative footprint, i.e. the sum of the distribution and intensity/probability of stressors or valued components at a given location
}
