% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulfoot.R
\name{cumulfoot}
\alias{cumulfoot}
\title{Cumulative footprint}
\usage{
cumulfoot(dat)
}
\arguments{
\item{dat}{data.frame with stressors intensity or valued components, with two columns: \code{uid} for the unique identifier of the grid cells; \code{value} for the value of the stressors or valued components that should be cumulated}
}
\description{
Function to evaluate cumulative footprint, i.e. the sum of individual elements or their intensity at a given location, of stressors or valued components
}
