% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{quantNorm}
\alias{quantNorm}
\alias{logTrans}
\alias{dataTrans}
\title{Data normalization}
\usage{
quantNorm(x)

logTrans(x)

dataTrans(x)
}
\arguments{
\item{x}{numeric vector of data to transform}
}
\description{
Function to scale data between 0 and 1 using the 99th percentile

Wrapper function to log transform data. This function could/should become more complex eventually

Function to normalize and transform data for further analyses
}
\examples{
dat <- runif(10, 0, 100)
x <- quantNorm(dat)
cbind(dat,x)
dat <- runif(10, 0, 100)
x <- logTrans(dat)
cbind(dat,x)
dat <- runif(10, 0, 100)
x <- dataTrans(dat)
cbind(dat,x)
}
