% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_cea.R
\name{direct_cea}
\alias{direct_cea}
\title{Cumulative effects assessments - Direct effects module}
\usage{
direct_cea(drivers, vc, sensitivity)
}
\arguments{
\item{drivers}{list of stars or terra objects}

\item{vc}{list of stars or terra objects}

\item{sensitivity}{matrix of \link{drivers x vc}}
}
\description{
Function to assess cumulative effects considering only direct effects, i.e. the Halpern et al. (2008) methodology
}
