% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{export_result}
\alias{export_result}
\title{Export rcea_result to raster files and metadata}
\usage{
export_result(
  result,
  dir,
  prefix = "result",
  overwrite = FALSE,
  metadata = "metadata.csv"
)
}
\arguments{
\item{result}{rcea_result}

\item{dir}{output directory}

\item{prefix}{filename prefix}

\item{overwrite}{logical, overwrite existing files}

\item{metadata}{name of metadata file (csv)}
}
\value{
invisible list with paths
}
\description{
Export rcea_result to raster files and metadata
}
