% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{exposure_cube}
\alias{exposure_cube}
\title{Run exposure from a cube}
\usage{
exposure_cube(cube, ...)
}
\arguments{
\item{cube}{rcea_cube with \code{stack$drivers} and \code{stack$vc}.}

\item{...}{Additional arguments passed to \code{exposure()}.}
}
\description{
Convenience wrapper around \code{exposure()} that uses stacks stored in
an \code{rcea_cube}.
}
\examples{
layers_path <- system.file("extdata/catalog/layers.csv", package = "rcea")
groups_path <- system.file("extdata/catalog/groups.yaml", package = "rcea")
catalog <- load_catalog(layers_path, groups_path)
cube <- make_cube(catalog)
cube <- stack_layers(cube)
expo <- exposure_cube(cube)
expo

}
