% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncea_species.R
\name{get_direct_indirect_sp}
\alias{get_direct_indirect_sp}
\title{========================================================================================

\if{html}{\out{<div class="sourceCode">}}\preformatted{----------------------------------------------------------------------------------------
}\if{html}{\out{</div>}}}
\usage{
get_direct_indirect_sp(motif_effects)
}
\description{
========================================================================================

\if{html}{\out{<div class="sourceCode">}}\preformatted{----------------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
