% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_array}
\alias{make_array}
\alias{make_stars}
\title{\code{rcea} helper functions}
\usage{
make_array(dat)

make_stars(dat, drivers, vc)
}
\arguments{
\item{dat}{list of cea matrices or cea array}

\item{drivers}{distribution and intensity of environmental drivers as stars object}

\item{vc}{distribution of valued components as stars object}
}
\description{
List of functions that support the main functions of the \code{rcea} package
}
\section{Functions}{
\itemize{
\item \code{make_array()}: create array from list of cea matrices

\item \code{make_stars()}: create stars object from list of cea matrices or cea array

}}
