% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-cea.R
\docType{package}
\name{learnCEA}
\alias{learnCEA}
\title{learnCEA: Learn Cost-effectiveness Modeling with R}
\description{
Course materials for learning how to perform applied cost-effectiveness analysis with R.
}
