% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{make_result}
\alias{make_result}
\title{Construct an rcea result object}
\usage{
make_result(data, layer_map, meta = list())
}
\arguments{
\item{data}{SpatRaster or matrix with template attribute.}

\item{layer_map}{data.frame with columns: layer, vc, driver.}

\item{meta}{list of metadata.}
}
\value{
rcea_result object.
}
\description{
Construct an rcea result object
}
\keyword{internal}
