% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncea.R
\name{ncea}
\alias{ncea}
\alias{cea_binary}
\alias{triads}
\alias{cea_pathways}
\alias{ncea_pathways}
\alias{ncea_pathways_}
\alias{ncea_motifs}
\alias{ncea_effects}
\alias{get_species_contribution}
\alias{get_direct_indirect}
\alias{get_net}
\alias{get_cekm_ncea}
\alias{ncea_split}
\title{Network-scale cumulative effects assessments}
\usage{
ncea(
  drivers,
  vc,
  sensitivity,
  metaweb,
  trophic_sensitivity,
  w_d = 0.5,
  w_i = 0.25,
  exportAs = "stars"
)

cea_binary(effect)

triads(metaweb, trophic_sensitivity)

cea_pathways(effect, vc)

ncea_pathways(vc_id, motifs)

ncea_pathways_(direct_pathways, motifs)

ncea_motifs(direct_effect, indirect_pathways)

ncea_effects(motif_summary, w_d = 0.5, w_i = 0.25)

get_species_contribution(motif_effects)

get_direct_indirect(motif_effects)

get_net(motif_effects)

get_cekm_ncea(motif_effects, vc)

ncea_split(
  drivers,
  vc,
  sensitivity,
  metaweb,
  trophic_sensitivity,
  w_d = 0.5,
  w_i = 0.25,
  output = "output/ncea",
  niter = NULL,
  run = NULL
)
}
\arguments{
\item{drivers}{distribution and intensity of environmental drivers as stars object}

\item{vc}{distribution of valued components as stars object}

\item{sensitivity}{matrix of environmental drivers and valued component, with same name as those used in \code{drivers} and \code{vc}}

\item{metaweb}{matrix of valued component by valued component describing the binary interations structuring the network of valued components}

\item{trophic_sensitivity}{data.frame of trophic sensitivities, default from Beauchesne. Available as data package with \code{data(trophic_sensitivity)}}

\item{exportAs}{string, the type of object that should be created, either multiple "data.frame" or "stars" objects}
}
\description{
Assessment of cumulative effects and related metrics using the Beauchesne et al. 2021 method.
}
\section{Functions}{
\itemize{
\item \code{cea_binary()}: transform effects assessment into binary 2D matrix to assess the presence of an effect to a valued component in a specific grid cell

\item \code{triads()}: assess all triads of interest from metaweb and attach trophic sensitivities

\item \code{cea_pathways()}: pathways of direct effect

\item \code{ncea_pathways()}: assess all triads of interest from metaweb

\item \code{ncea_pathways_()}: apply ncea_pathways to get pathways of indirect effect and trophic sensitivity for all cells

\item \code{ncea_motifs()}: get effects of drivers for species in all motifs in each cell

\item \code{ncea_effects()}: evaluate effects for all motifs using trophic sensitivity and effect weights

\item \code{get_species_contribution()}: get contribution of species to indirect effects

\item \code{get_direct_indirect()}: get direct and indirect effects of drivers

\item \code{get_net()}: get net effects of drivers

\item \code{get_cekm_ncea()}: get effects per km2

\item \code{ncea_split()}: split the assessment in smaller parts for larger analyses that run into memory issues and need to be run in parallel

}}
\examples{
# Data
drivers <- rcea:::drivers 
vc <- rcea:::vc
sensitivity <- rcea:::sensitivity
metaweb <- rcea:::metaweb
trophic_sensitivity <- rcea::trophic_sensitivity

\dontrun{
# Network-scale effects 
beauchesne <- ncea(drivers, vc, sensitivity, metaweb, trophic_sensitivity)
plot(beauchesne$net)
plot(beauchesne$direct)
plot(beauchesne$indirect)
}
}
