% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncea_.R
\name{ncea_}
\alias{ncea_}
\title{Network-scale cumulative effects assessments}
\usage{
ncea_(
  drivers,
  vc,
  sensitivity,
  metaweb,
  trophic_sensitivity,
  w_d = 0.5,
  w_i = 0.25,
  vc_focus,
  direct_effect = NULL,
  motif_summary = NULL
)
}
\arguments{
\item{drivers}{distribution and intensity of environmental drivers as stars object}

\item{vc}{distribution of valued components as stars object}

\item{sensitivity}{matrix of environmental drivers and valued component, with same name as those used in \code{drivers} and \code{vc}}

\item{metaweb}{matrix of valued component by valued component describing the binary interations structuring the network of valued components}

\item{trophic_sensitivity}{data.frame of trophic sensitivities, default from Beauchesne. Available as data package with \code{data(trophic_sensitivity)}}

\item{motif_summary}{list of parameters obtained from \link{ncea_motifs} to allow for faster assessments, since that part of the assessment is the longest to run.}

\item{exportAs}{string, the type of object that should be created, either a "data.frame" or a "stars" object}
}
\description{
Assessment of cumulative effects and related metrics using the Beauchesne et al. 2021 method.
}
\section{Functions}{
\itemize{
\item \code{ncea_()}: effects assessment using the Beauchesne approach, i.e. assessment of direct and indirect effects of drivers on valued components (species), with method applied on a single species at a time for ease of parallelisation

}}
\examples{
# Data
drivers <- rcea:::drivers 
vc <- rcea:::vc
sensitivity <- rcea:::sensitivity
metaweb <- rcea:::metaweb
data(trophic_sensitivity)

\dontrun{
# Network-scale effects 
beauchesne <- ncea(drivers, vc, sensitivity, metaweb, trophic_sensitivity)
plot(beauchesne$net)
plot(beauchesne$direct)
plot(beauchesne$indirect)
}
}
