% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_assessment.R
\name{new_assessment}
\alias{new_assessment}
\title{Creates new files from templates}
\usage{
new_assessment(name = "assessment")
}
\arguments{
\item{name}{name of the data workflow; defaults to \code{assessment}}
}
\value{
This function creates new files from the template available in \verb{inst/templates/}
}
\description{
This function is used to create new files from templates available in the \code{rcea} package.
}
\examples{
\dontrun{
# New assessment workflow
new_assessment()
}
}
