% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{normalize_aoi}
\alias{normalize_aoi}
\title{Normalize AOI input}
\usage{
normalize_aoi(aoi)
}
\arguments{
\item{aoi}{sf object, bbox list with xmin/xmax/ymin/ymax/crs, or WKT string}
}
\value{
SpatVector
}
\description{
Accepts sf object, bbox list, or WKT string and returns a terra SpatVector.
}
\examples{
bbox_path <- system.file("extdata/aoi/bbox.yml", package = "rcea")
bbox <- yaml::read_yaml(bbox_path)
aoi1 <- normalize_aoi(bbox)
aoi1

geojson_path <- system.file("extdata/aoi/aoi.geojson", package = "rcea")
aoi2 <- normalize_aoi(geojson_path)
aoi2
}
