% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Present value}
\usage{
pv(z, dr, t)
}
\arguments{
\item{z}{A numeric quantity.}

\item{dr}{The discount rate.}

\item{t}{A vector of times to compute the present value.}
}
\description{
Compute the present value of a numeric quantiy \code{z} given a discount rate
\code{dr} and times \code{t}.
}
\examples{
pv(1000, dr = .03, t = 0:4)
}
