% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rcea}
\alias{rcea-package}
\alias{rcea}
\title{rcea: package to perform cumulative effects assessments}
\description{
Package to perform cumulative effects assessments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ecosystem-assessments.github.io/rcea/}
  \item Report bugs at \url{https://ecosystem-assessments.github.io/rcea/issues}
}

}
\author{
\strong{Maintainer}: David Beauchesne \email{david.beauchesne@hotmail.com} (\href{https://orcid.org/0000-0002-3590-8161}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kevin Cazelles \email{kcazelle@uoguelph.ca} (\href{https://orcid.org/0000-0001-6619-9874}{ORCID})
}

}
