% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mstate3_data.R
\name{sim_mstate3_data}
\alias{sim_mstate3_data}
\title{Simulate example 3-state data}
\usage{
sim_mstate3_data(n = 2000, seed = NULL)
}
\arguments{
\item{n}{The number of patients to simulate.}

\item{seed}{An integer to pass to \code{set.seed()}. If \code{NULL},
then no seed is set.}
}
\value{
Returns a \code{data.table} with multiple rows for patient. That is,
for a given patient there are rows for each possible transition from a given
health state. The returned columns are as follows:
\describe{
\item{from}{The health state ID transitioned from.}
\item{to}{he health state ID transitioned to.}
\item{female}{\code{1} if female and \code{0} if male.}
\item{age}{The age of the patient in years.}
\item{patient_id}{The patient ID.}
\item{final}{An indicator equal to 1 if a patient is in their final health
state during the simulation and 0 otherwise.}
\item{time_start}{The time at the start of the interval.}
\item{time_stop}{The time and the end of the interval.}
\item{time}{The duration of the interval: \code{time_stop} - \code{time_start}.}
\item{status}{\code{1} if \code{time} is an observed transition and \code{0} if there was
right censoring.}
\item{transition_id}{The transition ID representing a unique transition from
\code{from} to \code{to}.}
\item{strategy_id}{The treatment strategy ID.}
\item{strategy_name}{The name of the treatment strategy. Either \verb{"New}
or \verb{"SOC}.}
}
}
\description{
Simulate example multi-state data for a 3-state healthy-sick-dead model
using Weibull distributions for each transition. Patients are
right censored after 15 years of followup.
}
\examples{
sim_mstate3_data(n = 3, seed = 101)
}
