% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mstate_data.R
\name{sim_mstate_data}
\alias{sim_mstate_data}
\title{Simulate example multi-state data}
\usage{
sim_mstate_data(n = 2000)
}
\arguments{
\item{n}{The number of patients to simulate.}
}
\description{
Simulate a 3-state healthy-sick-dead dataset with
right censoring using Weibull distributions for each transition.
}
