% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pfs_os.R
\name{sim_pfs_os_data}
\alias{sim_pfs_os_data}
\title{Simulate progression-free and overall survival}
\usage{
sim_pfs_os_data(n = 2000, seed = NULL)
}
\arguments{
\item{n}{The number of patients to simulate.}

\item{seed}{An integer to pass to \code{set.seed()}. If \code{NULL},
then no seed is set.}
}
\value{
Returns a \code{data.table} with one rows for each patient.
The returned columns are as follows:
\describe{
\item{patient_id}{The patient ID.}
\item{strategy_id}{The treatment strategy ID.}
\item{strategy_name}{The name of the treatment strategy. Either \verb{"New}
or \verb{"SOC}.}
\item{female}{\code{1} if female and \code{0} if male.}
\item{age}{The age of the patient in years.}
\item{pfs_time}{Time to progression, death, or right censoring.}
\item{pfs_status}{\code{1} if a progression or death event occured and \code{0}
if there was right censoring.}
\item{os_time}{Time to death or right censoring.}
\item{os_status}{\code{1} if a patient died and \code{0} if there was right censoring.}
}
}
\description{
Simulate a 3-state healthy-sick-dead dataset with
\code{sim_mstate3_date()} and compute progression-free survival (PFS)
and overall survival (OS).
}
\examples{
sim_pfs_os_data(n = 3, seed = 101)
}
