% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rcea.R
\name{use_rcea}
\alias{use_rcea}
\title{Use the \code{rcea} course materials}
\usage{
use_rcea(path = ".")
}
\arguments{
\item{path}{The path of the directory to put the course materials in.}
}
\description{
Add the \code{R} scripts that contain the code for the tutorials in the
\code{rcea} package to the directory specified by \code{path}.
}
\examples{
\dontrun{
usethis::create_project("~/Projects/rcea-exercises")
add_rcea("~/Projects/rcea-exercises")
}
}
