% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{exposure}
\alias{exposure}
\alias{cea}
\alias{ncea}
\alias{cea_array}
\alias{cea_stars}
\alias{cea_binary}
\alias{triads}
\alias{cea_pathways}
\alias{ncea_pathways}
\alias{ncea_pathways_}
\alias{ncea_motifs}
\alias{ncea_effects}
\alias{get_species_contribution}
\alias{get_direct_indirect}
\alias{get_net}
\alias{get_cekm}
\title{Cumulative effects assessments}
\usage{
exposure(drivers, vc, exportAs = "list")

cea(drivers, vc, sensitivity, exportAs = "list")

ncea(
  drivers,
  vc,
  sensitivity,
  metaweb,
  trophic_sensitivity,
  w_d = 0.5,
  w_i = 0.25,
  motif_summary = NULL
)

cea_array(dat)

cea_stars(dat, drivers, vc)

cea_binary(effect)

triads(metaweb)

cea_pathways(effect, vc)

ncea_pathways(vc_id, motifs)

ncea_pathways_(direct_pathways, motifs)

ncea_motifs(direct_effect, indirect_pathways)

ncea_effects(motif_summary, w_d = 0.5, w_i = 0.25)

get_species_contribution(motif_effects)

get_direct_indirect(motif_effects)

get_net(motif_effects)

get_cekm(motif_effects, vc)
}
\arguments{
\item{drivers}{distribution and intensity of environmental drivers as stars object}

\item{vc}{distribution of valued components as stars object}

\item{exportAs}{string, the type of object that should be created by \link{exposure} ("list" | "stars"), \link{cea} ("list" | "stars") and \link{ncea} ("data.frame" | "stars")}

\item{sensitivity}{matrix of environmental drivers and valued component, with same name as those used in \code{drivers} and \code{vc}}

\item{metaweb}{matrix of valued component by valued component describing the binary interations structuring the network of valued components}

\item{trophic_sensitivity}{data.frame of trophic sensitivities, default from Beauchesne. Available as data package with \code{data(trophic_sensitivity)}}

\item{motif_summary}{list of parameters obtained from \link{ncea_motifs} to allow for faster assessments, since that part of the assessment is the longest to run.}
}
\description{
Series of functions to assess cumulative effects and related metrics using the Halpern et al. 2008 or the Beauchesne et al. 2021 methodologies.
}
\section{Functions}{
\itemize{
\item \code{exposure()}: evaluate exposure of valued components to environmental drivers

\item \code{cea()}: effects assessment using the Halpern approach, i.e. assessment of direct effects of drivers on valued components

\item \code{ncea()}: assessment of effects across all motifs

\item \code{cea_array()}: create array from list of cea matrices

\item \code{cea_stars()}: create stars object from list of cea matrices or cea array

\item \code{cea_binary()}: transform effects assessment into binary 2D matrix to assess the presence of an effect to a valued component in a specific grid cell

\item \code{triads()}: assess all triads of interest from metaweb

\item \code{cea_pathways()}: pathways of direct effect

\item \code{ncea_pathways()}: pathways of indirect effect and trophic sensitivity

\item \code{ncea_pathways_()}: apply ncea_pathways to get pathways of indirect effect and trophic sensitivity for all cells

\item \code{ncea_motifs()}: get effects of drivers for species in all motifs in each cell

\item \code{ncea_effects()}: evaluate effects for all motifs using trophic sensitivity and effect weights

\item \code{get_species_contribution()}: get contribution of species to indirect effects

\item \code{get_direct_indirect()}: get direct and indirect effects of drivers

\item \code{get_net()}: get net effects of drivers

\item \code{get_cekm()}: get effects per km2

}}
\examples{
# Data
drivers <- rcea:::drivers 
vc <- rcea:::vc
sensitivity <- rcea:::sensitivity
metaweb <- rcea:::metaweb
data(trophic_sensitivity)

# Exposure
(expo <- exposure(drivers, vc, "stars"))
plot(expo)
expo <- merge(expo, name = "vc") |>
        split("drivers")
plot(expo)

# Species-scale effects
(halpern <- cea(drivers, vc, sensitivity, "stars"))
plot(halpern)
halpern <- merge(halpern, name = "vc") |>
          split("drivers")
plot(halpern)

\dontrun{
# Network-scale effects 
beauchesne <- ncea(drivers, vc, sensitivity, metaweb, trophic_sensitivity)
plot(beauchesne$net)
plot(beauchesne$direct)
plot(beauchesne$indirect)
}
}
