% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{cea}
\alias{cea}
\alias{get_cekm_cea}
\title{Cumulative effects assessments}
\usage{
cea(drivers, vc, sensitivity, exportAs = "list")

get_cekm_cea(dat, vc)
}
\arguments{
\item{drivers}{distribution and intensity of environmental drivers as stars object}

\item{vc}{distribution of valued components as stars object}

\item{sensitivity}{matrix of environmental drivers and valued component, with same name as those used in \code{drivers} and \code{vc}}

\item{exportAs}{string, the type of object that should be created, either a "list" or a "stars" object}
}
\description{
Assessment of cumulative effects using the Halpern et al. 2008 method.
}
\section{Functions}{
\itemize{
\item \code{get_cekm_cea()}: get effects per km2

}}
\examples{
# Data
drivers <- rcea:::drivers 
vc <- rcea:::vc
sensitivity <- rcea:::sensitivity

# Species-scale effects
(halpern <- cea(drivers, vc, sensitivity, "stars"))
plot(halpern)
halpern <- merge(halpern, name = "vc") |>
          split("drivers")
plot(halpern)
get_cekm_cea(halpern, vc)
}
