% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_extract.R
\name{cea_extract}
\alias{cea_extract}
\title{Extract attributes and aggregate cumulative effects assessment results}
\usage{
cea_extract(dat, dr_sel = NULL, vc_sel = NULL, cumul_fun = "none")
}
\arguments{
\item{dat}{stars object, either the drivers, valued components, exposure, cumulative effects assessment or network-sacale cumulative effects assessment results}

\item{dr_sel}{string, name of drivers to extract}

\item{vc_sel}{string, name of valued components to extract}

\item{cumul_fun}{function to apply on the stars object, one of "drivers" for the cumulative effects of each drivers on all value components, "vc" for the cumulative effects of all drivers on each valued component, "full" for the cumulative effects of all drivers on all valued components, "footprint" for the cumulative footprint of drivers or valued components, and "none" to keep data as is, i.e. to extract specific drivers and valued components from the assessment results.}
}
\description{
Function to extract specific drivers and vc of interest from the drivers, valued component, exposure and cumulative effects assessments results, and to aggregate data over drivers, valued components, or drivers and valued components.
}
\examples{
# Data
drivers <- rcea:::drivers 
vc <- rcea:::vc
sensitivity <- rcea:::sensitivity

# Cumulative effects assessment
dat <- cea(drivers, vc, sensitivity, "stars")

# Extract attributes
dr_sel <- c("driver1","driver5")
vc_sel <- c("vc4","vc7","vc10","vc12")
cea_extract(dat, dr_sel = dr_sel, vc_sel = vc_sel) 

# Cumulative footprint of selected drivers and valued components
cea_extract(drivers, dr_sel = dr_sel, cumul_fun = "footprint") 
cea_extract(vc, vc_sel = vc_sel, cumul_fun = "footprint") 

# Cumulative effects of all drivers on all vc
cea_extract(dat, cumul_fun = "drivers")

# Cumulative effects of all drivers on each vc
cea_extract(dat, cumul_fun = "vc") 

# Full cumulative effects
cea_extract(dat, cumul_fun = "full") 

}
