% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{collect_layers}
\alias{collect_layers}
\title{Materialize stack to memory or return SpatRaster}
\usage{
collect_layers(cube, collect = FALSE)
}
\arguments{
\item{cube}{rcea_cube}

\item{collect}{Logical, if TRUE returns in-memory copy}
}
\description{
Materialize stack to memory or return SpatRaster
}
