% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{exposure}
\alias{exposure}
\title{Cumulative exposure assessment}
\usage{
exposure(drivers, vc, exportAs = "list")
}
\arguments{
\item{drivers}{distribution and intensity of environmental drivers as stars object}

\item{vc}{distribution of valued components as stars object}

\item{exportAs}{string, the type of object that should be created, either a "list" or a "stars" object}
}
\description{
Assessment of the exposure (i.e. overlap) between valued components and environmental drivers in the context of cumulative effects assessments
}
\examples{
# Data
drivers <- rcea:::drivers 
vc <- rcea:::vc

# Exposure
(expo <- exposure(drivers, vc, "stars"))
plot(expo)
expo <- merge(expo, name = "vc") |>
        split("drivers")
plot(expo)
}
