% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-cea.R
\docType{package}
\name{learnCEA}
\alias{learnCEA}
\title{learnCEA: Learn Cost-effectiveness Modeling with R}
\description{
A companion package for the Professional Society for Health Economics
and Outcomes Research (ISPOR) short course "Applied cost-effectiveness modeling with R".
}
