% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{load_catalog}
\alias{load_catalog}
\title{Load and validate catalog}
\usage{
load_catalog(
  layers = file.path("catalog", "layers.csv"),
  groups = file.path("catalog", "groups.yaml")
)
}
\arguments{
\item{layers}{Path to layers.csv}

\item{groups}{Path to groups.yaml}
}
\value{
A list with \code{layers} tibble and \code{groups} list.
}
\description{
Load and validate catalog
}
