% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{make_cube}
\alias{make_cube}
\title{Cube object constructor}
\usage{
make_cube(catalog, config, aoi = NULL)
}
\arguments{
\item{catalog}{List with layers/groups from \code{load_catalog()}}

\item{config}{List from \code{load_config()}}

\item{aoi}{Optional AOI (sf/bbox/WKT/SpatVector)}
}
\value{
An object of class \code{rcea_cube}
}
\description{
Cube object constructor
}
