% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcea-package.R
\docType{package}
\name{rcea-package}
\alias{rcea}
\alias{rcea-package}
\title{rcea: Cumulative effects assessments package}
\description{
Package providing functionalities to perform spatially explicit cumulative effects assessments
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/david-beauchesne/rcea}
  \item Report bugs at \url{https://github.com/david-beauchesne/rcea/issues}
}

}
\author{
\strong{Maintainer}: David Beauchesne \email{david.beauchesne@hotmail.com} (\href{https://orcid.org/0000-0002-3590-8161}{ORCID}) [copyright holder]

}
\keyword{internal}
