% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{stack_layers}
\alias{stack_layers}
\title{Stack layers into a SpatRaster}
\usage{
stack_layers(cube, layer_ids = NULL, cache = cube$config$cache$use_cache)
}
\arguments{
\item{cube}{rcea_cube}

\item{layer_ids}{Optional vector of layer_ids to stack (default all)}

\item{cache}{Use cache (logical)}
}
\value{
SpatRaster
}
\description{
Stack layers into a SpatRaster
}
